Organize ingest package
