+++ app-backend/ingest/src/main/scala/com/azavea/rf/ingest/model/ImageSource.scala
@@ -1,4 +1,4 @@
-package com.azavea.rf.ingest
+package com.azavea.rf.ingest.model
 
 import spray.json._
 import DefaultJsonProtocol._
