+++ app-backend/ingest/src/main/scala/com/azavea/rf/ingest/model/Scene.scala
@@ -1,4 +1,4 @@
-package com.azavea.rf.ingest
+package com.azavea.rf.ingest.model
 
 import spray.json._
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
