+++ app-backend/ingest/src/test/scala/com/azavea/rf/ingest/model/JobDefinitionSpec.scala
@@ -1,4 +1,4 @@
-package com.azavea.rf.ingest
+package com.azavea.rf.ingest.model
 
 import spray.json._
 import org.scalatest._
